refactor(trainer-web): 스케줄·리포트 페이지의 위젯 분리
한 파일이 페이지 조립·상태·위젯·다이얼로그를 모두 들고 있어, 서로 무관한
작업이 같은 자리에서 충돌했다. 다른 기능이 이미 쓰는 배치대로 역할별로
나눈다.

- 스케줄: 캘린더(날짜 행·주 스트립·주 그리드·타임라인 행), 세션 카드와
  상태·유형 칩, 프로그램·관리·메모 블록, 세션 시트와 완료·취소 다이얼로그,
  반복 미리보기, 상담 인박스 액션을 presentation/widgets 로 옮겼다.
- 프로그램 초안(ProgramDraft)은 입력 컨트롤러 묶음이라 domain 이 아니라
  presentation/models 로 옮겼다.
- 리포트: 고객 선택기, 고객 리포트 본문, 지표 추이·4주 추이·주 비교,
  피드백 편집기, AI 요약 카드, 공유 메뉴, 일자별 상세, PDF 내보내기
  다이얼로그를 presentation/widgets 로 옮겼다.
- 두 페이지 파일에는 화면 조립과 상태만 남았다.

옮기면서 공개된 위젯에는 super.key 와 짧은 dartdoc 을 달았고, 잘못 붙어
있던 문서 한 줄(_ConsultationInboxAction 머리의 주 캘린더 설명)과 사라진
클래스를 가리키던 참조([_WeekNav])를 바로잡았다. 그 밖에는 순수 이동이라
동작은 그대로다.
