+++ frontend/flutter_trainer/lib/features/schedule/presentation/widgets/session_note_box.dart
@@ -1,0 +1,56 @@
+import 'package:flutter/material.dart';
+import 'package:oncare_trainer/design_system/tokens/colors.dart';
+import 'package:oncare_trainer/design_system/tokens/radius.dart';
+import 'package:oncare_trainer/design_system/tokens/spacing.dart';
+import 'package:oncare_trainer/gen/l10n/app_localizations.dart';
+
+/// 완료된 세션에 남긴 트레이너 메모 상자.
+class SessionNoteBox extends StatelessWidget {
+  const SessionNoteBox({super.key, required this.note});
+
+  final String note;
+
+  @override
+  Widget build(BuildContext context) {
+    final AppLocalizations l = AppLocalizations.of(context);
+    return Container(
+      padding: const EdgeInsets.symmetric(
+        horizontal: AppSpacing.md,
+        vertical: AppSpacing.sm,
+      ),
+      decoration: BoxDecoration(
+        color: AppColors.brandOrange.withValues(alpha: 0.06),
+        borderRadius: const BorderRadius.all(AppRadius.md),
+        border: Border(
+          left: BorderSide(
+            color: AppColors.brandOrange.withValues(alpha: 0.4),
+            width: 3,
+          ),
+        ),
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: <Widget>[
+          Text(
+            l.schedNote,
+            style: const TextStyle(
+              fontSize: 10,
+              fontWeight: FontWeight.w700,
+              // 메모지 표시다. 주의가 아니므로 빨강으로 올리지 않는다(#690).
+              color: AppColors.brandOrange,
+            ),
+          ),
+          const SizedBox(height: 2),
+          Text(
+            note,
+            style: const TextStyle(
+              fontSize: 12,
+              fontWeight: FontWeight.w500,
+              color: AppColors.mutedForeground,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
